requirements.txt
python-telegram-bot==20.5
python-dotenv
requests
fastapi>=0.68.0
uvicorn==0.22.0
